+++ src/igor/igor/renderer/iRenderEngine.h
@@ -46,7 +46,7 @@ namespace igor
     {
     public:
         /*! does nothing
-        */
+         */
         iRenderEngine() = default;
 
         /*! does nothing
@@ -54,7 +54,7 @@ namespace igor
         virtual ~iRenderEngine() = default;
 
         /*! sets the scene to render
-        */
+         */
         void setScene(iEntityScenePtr scene);
 
         /*! sets the current camera
@@ -69,62 +69,64 @@ namespace igor
         */
         void addMesh(iEntityPtr mesh);
 
+        /*! setup camera for render
+
+        \param camera the camera to use for setup
+        \param viewport the viewport given by the parent view
+        */
+        void setupCamera(const iaRectanglei &viewport);
+
         /*! renders given data
 
         \param viewport the given viewport to render in
         */
-        void render(const iaRectanglei &viewport);
+        void render();
 
         /*! \returns current frustum
 
         valid after render/setupCamera
         */
-        const iFrustumd& getFrustum() const;
+        const iFrustumd &getFrustum() const;
 
     private:
-
         /*! the scene to render
-        */
+         */
         iEntityScenePtr _scene = nullptr;
 
         /*! camera ID
-        */
+         */
         iEntityID _cameraID = iEntityID::getInvalid();
 
         /*! current frustum
-        */
+         */
         iFrustumd _frustum;
 
         struct iInstaningPackage
         {
             iInstancingBufferPtr _buffer;
             iMaterialPtr _material; // TODO needs to be part of the buffer so individual instances can have different colors etc
-        };        
+        };
 
         /*! bringing all nodes using the same material together for more efficient rendering
-        */
+         */
         struct iMaterialGroup
         {
             /*! the shader used
-            */
+             */
             iShaderPtr _shader;
 
             /*! optional instancing buffers per mesh that is using the same material
-            */
+             */
             std::unordered_map<iMeshPtr, iInstaningPackage> _instancing;
         };
 
         /*! render nodes
          */
-        std::vector<iMaterialGroup> _materialGroups;        
+        std::vector<iMaterialGroup> _materialGroups;
 
-        /*! setup camera for render
-
-        \param camera the camera to use for setup
-        \param viewport the viewport given by the parent view
-        */
-        void setupCamera(iEntityPtr camera, const iaRectanglei &viewport);
-
+        /*! renders all mesh instances
+         */
+        void renderInstances();
     };
 } // namespace igor